Get rid of garcon_{init,shutdown}.
Most likely garcon will only be used in normal applications,
so calling g_type_init() is handled by gtk_init(), if not
the user should call it.

For now an application implementing garcon should call
garcon_set_environment (...) before using any of the functions.
This will leak the string once, so if you want to avoid this
call garcon_set_environment (NULL) when the app exits..