Disable the parent call for getting the appdirs.
A submenu always inherits the application directories of its parent,
that is the reason the parent call was added.
It only turned out we were looking in that same directories for
.desktop files multiple times.

This was caused by the combination of the parent call and
traversing the children in garcon_menu_collect_files(). For each
submenu the appdirs of the root were added and traversed again.

This is not needed because we always start at the root and traverse
in "pre-order", so all the desktop files are added in the hash-table.