Release 0.1.0!
This is the first development release of garcon. Even though it is still
lacking menu monitoring support, it might be worth adopting it some
applications already. After all, we can always fall back to the old,
less fine grained monitoring if we don't have something ready in time
for Xfce 4.8.