Add garcon_menu_directory_load(). Further work on monitoring.
Require GarconMenuDirectories to be created and loaded in two steps (one
for creating the object, another one for loading), just like we do it
for GarconMenu.

Make it possible to reload GarconMenus with garcon_menu_load() without
creating a new GarconMenu before (this was required in libxfce4menu but
it is inconvenient). Split the monitoring initialization into several
methods, add TODOs for the missing pieces: reacting on app dir changes.