Additional work on app dir monitoring.
This adds new functionality to find the GarconMenuItem and all
GarconMenus corresponding to a GFile. App dirs are now only monitored by
the root menu.

We properly react on changes that cause a single desktop file to be
unreadable or to break parsing. Reacting on a deleted desktop file is
also partly support now (what still needs to be implemented is looking
for a desktop file with the same desktop-file ID, replacing the deleted
desktop file).

TODOs in garcon-menu.c were updated.

New signals introduced in this commit:
  - GarconMenu::item-added
  - GarconMenu::item-changed
  - GarconMenu::item-removed

New methods introduced in this commit:
  - garcon_menu_element_equal()
  - garcon_menu_item_cache_remove_file()
  - garcon_menu_item_pool_remove()