React on app dir changes which, in most cases, requires a menu reload.
Add a boolean affects_the_outside return parameter to
garcon_menu_item_reload() and garcon_menu_item_reload_from_file() and
set it to TRUE if the categories are different before and after
reloading the file.

Most app dir changes result in the emission of a "reload-required"
signal of the root menu, except for when a change is made to a .desktop
file that only affects the item itself, not its position in the menu
hierarchy.

The monitoring implement is very similar to the basic monitoring
algorithm we had in libxfce4menu, except that it's a little bit smarter
and takes merged files and menu file priorities into account. So a
reload can sometimes be avoided without handling all the different
scenarios that would have to be handled ideally.