Add garcon_menu_item_cache_invalidate_file, simplify app dir monitoring.
Only listen to G_FILE_MONITOR_EVENT_CHANGES_DONE_HINT events, otherwise
consecutive events may not be picked up. A good example is when a
desktop entry is edited and saved in VIM. This results in a series of
events (move, create, change etc.). Some of these will not be picked up
if garcon is busy reloading the menu at the time of their emission.

So, the only real choice we have is to hope that CHANGES_DONE_HINT is
always generated at the end of modifications like move, create, delete
and attribute changes.