Do not decrement the internal refcount of items.
This reverts commit c63c7b91f1b9cca982a9441de5397d4019fa6fb0. The
internal reference counter is used to check whether items are allocated
or not. Whenever an item has been included in a menu at least once (that
is, it matched an <Include> rule) it's refcount is not supposed to be
lowered anymore (in that sense it's more like a boolean flag than a
counter). Decrementing the refcount breaks <OnlyUnallocated/>.

Reverting this change might break menu reloading, so perhaps we need to
use something different than garcon_menu_item_unref() in
XfceMenuItemCache. Will check later.