Fix menu icons for misconfigured desktop files (Bug 10709)
The desktop entry spec says the Icon key in a .desktop file should
either not have an extension or provide an absolute path. This patch
allows programs which don't conform to that spec to still properly
display an icon by checking for the icon name before the extension
or if they exist in the pixmaps folder; Same as xfdesktop.
This patch also scales all the images down to the configured gtk icon
size which is slightly different from xfdesktop because garcon doesn't
depend on libexo.
Revelant Desktop Entry Spec:
http://standards.freedesktop.org/desktop-entry-spec/latest/ar01s05.html