Port Garcon to Gtk3
This adds Gtk3 support. Currently, both Gtk2 and Gtk3 are built
at the same time until the consumers of Garcon have all been
ported over. It uses the same codebase for both to assist with
troubleshooting issues. Note, this means the GtkImageMenuItems
are now an image + label packed into a box, so the padding
increased slightly.