garcon-gtk: Populate GtkMenus only when shown
When the display of menu item icons is required, the creation of the
entire application menu, typically, can take several tenths of a second
or even more than a second on low performance machines. This change
reduces the latency of the menu display, spreading the load over the
display of its submenus and thus doing the rest of the work only if
necessary.

Some of the code in `garcon_gtk_menu_add()` needs to be duplicated to
test for displayable items in a submenu without entering a recursive
process. If a submenu has displayable items, it is added as a menu item,
but is not populated until later, if its display is required.

Unfortunately, for reasons internal to GTK, this process is not easily
parallelized in a GTask like loading the GarconMenu in the next commit.
This is why we have no choice but to delay the task by executing it only
when necessary.

Starts !18.