garcon-gtk: Populate submenus only when shown (Fixes #5)
Some of the code in `garcon_gtk_menu_add()` needs to be duplicated to
test for displayable items in a submenu without entering a recursive
process. If a submenu has displayable items, it is added as a menu item,
but is not populated until later, if its display is required.

See also xfce4-panel#485.