garcon-gtk: Load GarconMenu asynchronously
This loading makes many disk accesses when reading the `.desktop` files,
which can take several seconds for only a hundred entries, on machines
equipped with mechanical disks and when these files have not yet been
cached.

This way, the loading is done as soon as possible in parallel, and we
wait until it is finished if necessary during a request to show the
menu. We can then start populating the GtkMenu synchronously, according
to the process described in the previous commit message.

Helps: #5, xfce4-panel#485
Fixes: #3, xfce4-panel#184
MR: !18