garcon-menu-separator: Fix constructor
This partially reverts 1045e3f045aa819dc065682943865c8d12fdb734, where
the static qualifier was removed without reason and without making the
rest of the function consistent with this change. The
GarconMenuSeparator is not intended to be packed in a container, so it
can effectively be created once and then only referenced.

This also fixes a memory leak reported by Valgrind on the weak reference
added on `separator`, which did not make sense anymore with a non-static
variable.