Wait for any async operation to finish before releasing locks
As the doc states, calling `g_mutex_clear()` on a locked mutex leads to
undefined behavior. This can happen at least in the case of GarconMenu,
where a load may be in progress at the time the last reference is
removed in GarconGtkMenu. This can be reproduced by reloading the panel
configuration via xfce4-panel-profiles when a custom menu file is used.

Fixes: 966f7c99dc5ebf39ec423883652b805b749a9a2d
Related: !18