garcon-gtk: Add missing lock when filling the GtkMenu
Another special case not properly locked: this filling can start while
the GarconMenu is loading, because its status is updated only at the end
of the loading.

The other occurrence of `garcon_gtk_menu_add()` that fills the root menu
is protected because it is in the GAsyncReadyCallback, so after the task
is finished and before another one is started. As for the third
occurrence of `garcon_gtk_menu_add()` which fills sub-menus, it is
protected by `g_signal_connect_object()` which keeps a reference on the
sub-GarconMenu.

Fixes criticals such as
"g_node_last_child: assertion 'node != NULL' failed".

Fixes: d2c7068e1d71738b5952f9a8823112709a4a56c4
Related: !18