garcon-gtk: Avoid populating a wrong menu
Between the end of load_async() and the start of load_finish(), one or
more calls may be made to set_menu(), or the garcon menu may be
destroyed and recreated by some other means. The menu may then be marked
as loaded when it is no longer, and any attempt to access it may lead to
critical warnings such as:
g_node_last_child: assertion 'node != NULL' failed

This can be reproduced, for example, by keeping the app menu shown when
reloading the panel configuration via xfce4-panel-profiles. To avoid
this, we ensure that the menu is the same at the input and output of the
asynchronous task by means of a pointer comparison.

Fixes: d2c7068e1d71738b5952f9a8823112709a4a56c4
Related: !18
(cherry picked from commit af63e00f315801ccdaf83526222876fc0f755455)